matcha-stock/scrapers/requirements.txt
playwright==1.32.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
requests==2.29.0 